requirements.txt
Flask
playwright
beautifulsoup4
requests
rich
google-cloud-storage
google-cloud-firestore
pyjwt
rembg
loguru
Pillow
numpy